requirements.txt
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
pathlib==1.0.1
pycodestyle==2.5.0
pycparser==2.19
python-dotenv==0.12.0
python-telegram-bot==12.4.2
six==1.14.0
tornado==6.0.3
